src/main/java/com/thoughtworks/springbootemployee/mapper/CompanyRequestMapper.java
package com.thoughtworks.springbootemployee.mapper;

import com.thoughtworks.springbootemployee.dto.CompanyRequest;
import com.thoughtworks.springbootemployee.model.Company;
import org.springframework.beans.BeanUtils;
import org.springframework.stereotype.Component;

@Component
public class CompanyRequestMapper {
    public Company toCompany(CompanyRequest companyRequest) {
        Company company = new Company();
        company.setCompanyName(companyRequest.getCompanyName());
        company.setEmployeesNumber(companyRequest.getEmployeesNumber());
        return company;
    }
}
